refactor: update import paths and remove deprecated utility files
- Changed import paths for various components and utilities to reflect the new structure, including updates to Firebase service and validation utilities.
- Removed unused utility files such as dataFormatter, validate, shortText, scrollToSection, and their associated imports across components.
- Ensured all affected components maintain functionality after the changes.
