+++ src/utils/date.js
@@ -1,6 +1,6 @@
 import { Timestamp } from "firebase/firestore";
 
-export const dataFormatter = (timestamp) => {
+export const dateFormatter = (timestamp) => {
   if (!timestamp) return ""; // Jeśli brak timestamp, zwróć pusty string
   const date =
     timestamp instanceof Timestamp ? timestamp.toDate() : new Date(timestamp);
